refactor: create separate components for backbtn & empty playlist page & change placement for files
