+++ src/components/PlaylistsPage/PlaylistEmptyPage/PlaylistEmptyPage.tsx
@@ -1,0 +1,4 @@
+const PlaylistEmptyPage = (): JSX.Element => {
+  return <div className="text-gray-400 text-center">No words in this playlist yet.</div>
+}
+export default PlaylistEmptyPage
