+++ src/routes.tsx
@@ -1,9 +1,9 @@
 import UserProfile from '@components/UserProfile/UserProfile.tsx'
-import PlaylistsPage from '@components/PlaylistPage/PlaylistsPage.tsx'
-import PlaylistDetailsPage from '@components/PlaylistPage/PlaylistDetailsPage.tsx'
+import PlaylistsPage from '@components/PlaylistsPage/PlaylistsPage.tsx'
+import PlaylistPage from '@components/PlaylistsPage/PlaylistPage/PlaylistPage.tsx'
 
 export const routes = [
   { path: '/', element: <PlaylistsPage /> },
   { path: '/profile', element: <UserProfile /> },
-  { path: '/playlist/:id', element: <PlaylistDetailsPage /> },
+  { path: '/playlist/:id', element: <PlaylistPage /> },
 ]
